requirements.txt
Flask==3.1.1
google-api-python-client==2.179.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
scikit-learn==1.7.1
pandas==2.3.1
numpy==2.3.2
scipy==1.16.1
